requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
typing-extensions
python-multipart
email-validator
httpx
starlette
python-jose[cryptography]  # For JWT handling if needed
geopy  # For additional geolocation utilities
